requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
openai==1.30.1
requests==2.31.0
pyyaml==6.0.1
typing-extensions==4.11.0